Create TokenStore and fn to get token from request
The TokenStore protocol is now used inside of workflow to deal with creating and verifying tokens. It can also be used from the webapp using the workflow to extend and destroy tokens in conjunction with the token/from-request fn.

There is an in memory impl of the TokenStore protocol called MemTokenStore